+++ NEWS.md
@@ -2,11 +2,14 @@
 
  * use `saveRDS`, rather than `save`, local caching
 
+ * change `print.Rmonad` parameter `print_value` to `value`
+
 ## Bug fixes
 
  * fix expression capture in `as_monad`
 
  * make `index` the second positional argument of `has_value`, as with other accessors
+
 
 # rmonad 0.4.0